% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_init.R
\name{build_authorization_uri}
\alias{build_authorization_uri}
\title{Generate URI for the OAuth authorization code flow}
\usage{
build_authorization_uri(resource, tenant, app, password = NULL,
  username = NULL, ...,
  aad_host = "https://login.microsoftonline.com/", version = 1)
}
\arguments{
\item{resource, tenant, app}{See the corresponding arguments for \link{get_azure_token}.}

\item{password}{A client secret to be sent to the authorization endpoint, if the app requires it. Note that this is \emph{not} your personal account password.}

\item{username}{An optional login hint to be sent to the authorization endpoint.}

\item{...}{Named arguments that will be added to the authorization URI as query parameters.}

\item{aad_host}{The base URL for the Azure Active Directory login endpoint.}

\item{version}{The AAD version, either 1 or 2.}
}
\value{
The authorization URI as a string.
}
\description{
Generate URI for the OAuth authorization code flow
}
\details{
This function is mainly for use in embedded scenarios, such as within a Shiny web app. In this case, the first stage of the authorization code flow (obtaining the code) must be handled separately from the second stage (using the code to obtain the token). The \code{get_authorization_uri} function returns the URI for obtaining the code, which can be set as a redirect from within a Shiny app's UI component.
}
