% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_auth_header}
\alias{format_auth_header}
\title{Format an AzureToken object}
\usage{
format_auth_header(token)
}
\arguments{
\item{token}{An Azure OAuth token.}
}
\description{
Format an AzureToken object
}
