% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_init.R
\name{request_device_code}
\alias{request_device_code}
\title{Initialize the OAuth device code flow}
\usage{
request_device_code(resource, tenant, app,
  aad_host = "https://login.microsoftonline.com/", version = 1)
}
\arguments{
\item{resource, tenant, app, aad_host, version}{See the corresponding arguments for \link{get_azure_token}.}
}
\value{
A list, containing the following components:
\itemize{
\item \code{user_code}: A short string to be shown to the user
\item \code{device_code}: A long string to verify the session with the AAD server
\item \code{verification_uri}: The URI the user should browse to in order to login
\item \code{expires_in}: The duration in seconds for which the user and device codes are valid
\item \code{interval}: The interval between polling requests to the AAD token endpoint
\item \code{message}: A string with login instructions for the user
}
}
\description{
Initialize the OAuth device code flow
}
\details{
This function is mainly for use in embedded scenarios, such as within a Shiny web app. In this case, the first stage of the device code flow (obtaining the code and displaying it to the user) must be handled separately from the second stage (polling the endpoint for the token). The \code{request_device_code} function returns a list containing the user and authorization codes, plus a message to be displayed to the user.
}
