% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uris.R
\name{authorization_uri}
\alias{authorization_uri}
\title{Generate URI for the OAuth authorization code flow}
\usage{
authorization_uri(endpoint, ...)
}
\arguments{
\item{endpoint}{The base URL for the Azure Active Directory login endpoint, or an \code{aad_endpoint} object.}

\item{...}{Named arguments that will be added to the authorization URI as query parameters.}

\item{tenant}{Your tenant. This can be a name ("myaadtenant"), a fully qualified domain name ("myaadtenant.onmicrosoft.com" or "mycompanyname.com"), or a GUID.}

\item{app}{The client/app ID to use to authenticate with.}

\item{version}{The AAD version, either 1 or 2.}
}
\value{
An object of class \code{httr::url}, representing the \emph{parsed} authorization URI. You can call \code{httr::build_url} on this object to obtain the URI as a text string.
}
\description{
Generate URI for the OAuth authorization code flow
}
