% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuth.R
\name{AzureR_dir}
\alias{AzureR_dir}
\title{Data directory for AzureR packages}
\usage{
AzureR_dir()
}
\value{
A string containing the data directory.
}
\description{
Data directory for AzureR packages
}
\details{
AzureAuth can store authentication credentials and OAuth tokens in a user-specific directory, using the rappdirs package. On recent Windows versions, this will usually be in the location \code{C:\\Users\\(username)\\AppData\\Local\\AzureR}. On Unix/Linux, it will be in \code{~/.local/share/AzureR}, and on MacOS, it will be in \code{~/Library/Application Support/AzureR}. The working directory is not touched (which significantly lessens the risk of accidentally introducing cached tokens into source control).

On package startup, if this directory does not exist, AzureAuth will prompt you for permission to create it. It's recommended that you allow the directory to be created, as otherwise you will have to reauthenticate with Azure every time. Note that many cloud engineering tools, including the \href{https://docs.microsoft.com/en-us/cli/azure/?view=azure-cli-latest}{Azure CLI}, save authentication credentials in this way.
}
\seealso{
\link{get_azure_token}

\link[rappdirs:user_data_dir]{rappdirs::user_data_dir}
}
